% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{hlra}
\alias{hlra}
\title{Perform generic HLRA problem with inverted autoregressive weights.}
\usage{
hlra(
  series,
  r,
  L = default_L(series),
  ar_coefs = NULL,
  alpha = 0.1,
  debug = FALSE,
  envelope = unit_envelope(series),
  compensated = TRUE,
  set_seed = NULL,
  additional_pars = list()
)
}
\arguments{
\item{series}{Source time series, numeric vector or list of numeric vectors (may contain NA's)}

\item{r}{Desired rank.}

\item{L}{Window length.}

\item{ar_coefs}{Autoregression coefficients.}

\item{alpha}{Parameter for Cadzow weights search}

\item{debug}{Debug mode on/off switch.}

\item{envelope}{Optional noise envelope.}

\item{compensated}{Use Compensated Horner Scheme in MGN algorithm?}

\item{set_seed}{Seed function for deterministic SVD decomposition}

\item{additional_pars}{Additional parameters for inner optimizers.}
}
\value{
Object of class "hlra" containing "signal" field, which is result of optimization.
}
\description{
Perform generic HLRA problem with inverted autoregressive weights.
}
\examples{
gap <- 18
series <- c(1, numeric(gap), 1)
x <- hlra(series, r = 1)
plot(x$signal, type = "b")
}
