% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{hlra_ar}
\alias{hlra_ar}
\title{Perform generic HLRA problem with unknown inverted autoregressive weights.}
\usage{
hlra_ar(
  series,
  r,
  p = 1,
  L = default_L(series),
  alpha = 0.1,
  k = p * 4,
  ar_coefs_eps = 1e-07,
  initial_ar_coefs = NULL,
  debug = FALSE,
  envelope = unit_envelope(series),
  compensated = TRUE,
  set_seed = NULL,
  additional_pars = list()
)
}
\arguments{
\item{series}{Source time series, numeric vector or list of numeric vectors (may contain NA's)}

\item{r}{Desired rank.}

\item{p}{Order of AR process.}

\item{L}{Window length.}

\item{alpha}{Parameter for Cadzow weights search}

\item{k}{Maximum iteration of AR parameters estimation}

\item{initial_ar_coefs}{Initial autoregression coefficients.}

\item{debug}{Debug mode on/off switch.}

\item{envelope}{Optional noise envelope.}

\item{compensated}{Use Compensated Horner Scheme in MGN algorithm?}

\item{set_seed}{Seed function for deterministic SVD decomposition}

\item{additional_pars}{Additional parameters for inner optimizers.}

\item{ar_coefs_epsilon}{AR coefficients threshold}
}
\value{
Object of class "hlra" containing "signal" field, which is result of optimization.
}
\description{
Perform generic HLRA problem with unknown inverted autoregressive weights.
}
\examples{
library(Rssa)
data("USUnemployment")
series <- log(USUnemployment[, 2])
x <- hlra_ar(series, r = 9, p = 3, alpha = .8, initial_ar_coefs = c(.9))
plot(x$signal)
}
