% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highorderoptimize.R
\name{hlra_cadzow}
\alias{hlra_cadzow}
\title{Perform Cadzow iterations.}
\usage{
hlra_cadzow(series, r, L = default_L(series), left_diags = NULL,
  right_diags = NULL, debug = FALSE, set_seed = NULL,
  additional_pars = list())
}
\arguments{
\item{series}{Source time series, numeric vector or list of numeric vectors (without NA's)}

\item{r}{Desired rank.}

\item{L}{Window length.}

\item{left_diags}{Diagonals of left SVD weights matrix.}

\item{right_diags}{Diagonals of right SVD weights matrix.}

\item{debug}{Debug mode on/off switch.}

\item{set_seed}{Seed function for deterministic SVD decomposition}

\item{additional_pars}{Additional parameters for inner optimizers.}
}
\value{
Object of class "hlra" containing "signal" field, which is result of optimization.
}
\description{
Perform Cadzow iterations.
}
\examples{
x <- hlra_cadzow(1:20, r = 1)
plot(x$signal, type = "b")
}
