% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highorderoptimize.R
\name{hlra_mgn}
\alias{hlra_mgn}
\title{Perform Modified Gauss-Newton Algorithm.}
\usage{
hlra_mgn(series, initial_glrr, weights = NULL,
  weights_chol = hlra_mgn_default_chol(series), debug = FALSE,
  compensated = TRUE, additional_pars = list())
}
\arguments{
\item{series}{Source time series, numeric vector or list of numeric vectors (may contain NA's)}

\item{initial_glrr}{Initial GLRR, which is non-null numeric vector of length more than 1.}

\item{weights}{Optional weights matrix of class Matrix.}

\item{weights_chol}{Optional Cholesky decomposition of weights matrix, matrix of class Matrix}

\item{debug}{Debug mode on/off switch.}

\item{compensated}{Use Compensated Horner Scheme in MGN algorithm?}

\item{additional_pars}{Additional parameters for inner optimizers.}
}
\value{
Object of class "hlra" containing "signal" field, which is result of optimization.
}
\description{
Perform Modified Gauss-Newton Algorithm.
}
\examples{
gap <- 18
series <- c(1, numeric(gap), 1)
x <- hlra_mgn(series, c(.3, .5))
plot(x$signal, type = "b")
}
