% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highorderoptimize.R
\name{hlra_sylvester}
\alias{hlra_sylvester}
\title{Find GCD of two polynomials using Modified Gauss-Newton Algorithm (i.e. Sylvester LRA).}
\usage{
hlra_sylvester(polynoms, initial_gcd, poly_weights = NULL,
  debug = FALSE, compensated = TRUE, additional_pars = list())
}
\arguments{
\item{polynoms}{Source polynomials, numeric vector or list of numeric vectors}

\item{initial_gcd}{Initial GCD vector, which is non-null numeric vector of length more than 1.}

\item{poly_weights}{Optional vector of weight of each polynomial.}

\item{debug}{Debug mode on/off switch.}

\item{compensated}{Use Compensated Horner Scheme in MGN algorithm?}

\item{additional_pars}{Additional parameters for inner optimizers.}
}
\value{
Object of class "hlra_sylvester" containing "gcd" field, which is result of optimization.
}
\description{
Find GCD of two polynomials using Modified Gauss-Newton Algorithm (i.e. Sylvester LRA).
}
\examples{
x <- hlra_sylvester(list(c(1, -3, 3, -1), c(1, -2, 1)), c(.1, .5))
print(x$gcd)
}
