% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highorderoptimize.R
\name{hlra_tune}
\alias{hlra_tune}
\title{Tune parameters of HLRA problem.}
\usage{
hlra_tune(series, r_range = 1:15, p_range = 0:3,
  L = default_L(series), alpha = 0.1,
  envelope = unit_envelope(series), set_seed = NULL, cluster = NULL,
  initial_ar_coefs = NULL, additional_pars = list())
}
\arguments{
\item{series}{Source time series, numeric vector or list of numeric vectors (may contain NA's)}

\item{r_range}{Possible values of time series rank.}

\item{p_range}{Possible values of order of AR process.}

\item{L}{Window length.}

\item{alpha}{Parameter for Cadzow weights search}

\item{envelope}{Optional noise envelope.}

\item{set_seed}{Seed function for deterministic SVD decomposition}

\item{cluster}{Optional SNOW cluster for parallel computing}

\item{initial_ar_coefs}{Initial autoregression coefficients.}

\item{additional_pars}{Additional parameters for inner optimizers.}
}
\value{
Data frame of class "hlra_tune" containing Bayesian Information Criterion (BIC) for each model.
}
\description{
Tune parameters of HLRA problem.
}
\examples{
library(Rssa)
data("USUnemployment")
seedf <- function() set.seed(15)
series <- USUnemployment[, 2]
series <- series[!is.na(series)]
bic_data <- tune_hlra(series, r_range = 8:12, p_range = 0:3, alpha = .8, initial_ar_coefs = c(.9), set_seed = seedf)
plot(bic_data)
}
