% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highorderoptimize.R
\docType{package}
\name{rhlra}
\alias{rhlra}
\alias{rhlra-package}
\title{rhlra: Hankel Low-Rank Approximation for R package.}
\description{
TODO
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

