% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highorderoptimize.R
\name{hlra_sylvester}
\alias{hlra_sylvester}
\title{Find GCD of two polynomials using Modified Gauss-Newton Algorithm (i.e. Sylvester LRA).}
\usage{
hlra_sylvester(
  polynoms,
  r,
  initial_poly = NULL,
  poly_weights = NULL,
  alpha = 0.1,
  debug = FALSE,
  compensated = TRUE,
  additional_pars = list()
)
}
\arguments{
\item{polynoms}{Source polynomials, numeric vector or list of numeric vectors}

\item{r}{Expected order of GCD}

\item{initial_poly}{Initial approximation for GCD, which is non-null numeric vector of length more than 1.}

\item{poly_weights}{Optional vector of weight of each polynomial.}

\item{alpha}{Parameter for Cadzow weights search}

\item{debug}{Debug mode on/off switch.}

\item{compensated}{Use Compensated Horner Scheme in MGN algorithm?}

\item{additional_pars}{Additional parameters for inner optimizers.}
}
\value{
Object of class "hlra_sylvester" containing "gcd" field, which is result of optimization.
}
\description{
Find GCD of two polynomials using Modified Gauss-Newton Algorithm (i.e. Sylvester LRA).
}
\examples{
x <- hlra_sylvester(list(c(1, -3, 3, -1), c(1, -2, 1)), 1)
print(x$gcd)
}
