% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{package}
\name{rhlra}
\alias{rhlra-package}
\alias{rhlra}
\title{rhlra: Hankel Low-Rank Approximation for R package.}
\description{
This package provides tools for solving Hankel Structured Low-Rank Approximation problems
}
\section{See Also}{

\code{\link{hlra}}, \code{\link{hlra_tune}}, \code{\link{hlra_cadzow}}, \code{\link{hlra_mgn}}.
}

